feat: add new app "runs" for handling the Run model and domain logic for state transitions
