+++ backend/README.md
@@ -1,0 +1,11 @@
+# Databús
+
+Django apps:
+
+- `api`: API endpoints for Databús, including REST, WebSocket, and SSE.
+- `realtime_engine`: Real-time engine (Celery tasks) for handling dynamic updates and events.
+- `schedule_engine`: Schedule engine (Celery tasks) for handling scheduled updates and events.
+- `operations`: Operational data
+- `runs`: Run data (_anima machinae_)
+- `website`: Miscellaneous website features (e.g., admin interface, users, etc.)
+- `feed`: GTFS data handling and processing
