+++ backend/api/serializers.py
@@ -5,6 +5,8 @@ from operations.models import (
     Vehicle,
     Equipment,
     EquipmentLog,
+)
+from runs.models import (
     Run,
     Position,
     Progression,
