+++ backend/operations/admin.py
@@ -6,10 +6,6 @@ from .models import (
     Company,
     Equipment,
     EquipmentLog,
-    Run,
-    Position,
-    Progression,
-    Occupancy,
 )
 
 # Register your models here.
@@ -20,7 +16,3 @@ admin.site.register(DataProvider)
 admin.site.register(Company, admin.GISModelAdmin)
 admin.site.register(Equipment)
 admin.site.register(EquipmentLog)
-admin.site.register(Run)
-admin.site.register(Position, admin.GISModelAdmin)
-admin.site.register(Progression)
-admin.site.register(Occupancy)
